+++ src/main/java/com/practicasalma/proyectoalma/model/Persona.java
@@ -13,7 +13,7 @@ public abstract class Persona {
     // Hibernate autoincrementará este ID en cada tabla hija
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     // Obligatorios
     @Column(nullable = false)
@@ -62,7 +62,7 @@ public abstract class Persona {
 
     // Getters y Setters
 
-    public Integer getId() { return id; }
+    public Long getId() { return id; }
 
     public String getNombre() { return nombre; }
     public void setNombre(String nombre) { this.nombre = nombre; }
